conf: reorganize/split code to utils.c
Move run_script/run_script_argv helpers to utils.c
